Move authenticator instance out of listener file
