+++ proxy/listener.go
@@ -5,7 +5,6 @@ import (
 	"net"
 	"strconv"
 
-	"github.com/Dreamacro/clash/component/auth"
 	"github.com/Dreamacro/clash/proxy/http"
 	"github.com/Dreamacro/clash/proxy/redir"
 	"github.com/Dreamacro/clash/proxy/socks"
@@ -17,8 +16,6 @@ var (
 	socksListener *socks.SockListener
 	httpListener  *http.HttpListener
 	redirListener *redir.RedirListener
-
-	authenticator auth.Authenticator
 )
 
 type listener interface {
@@ -38,14 +35,6 @@ func AllowLan() bool {
 
 func SetAllowLan(al bool) {
 	allowLan = al
-}
-
-func Authenticator() auth.Authenticator {
-	return authenticator
-}
-
-func SetAuthenticator(au auth.Authenticator) {
-	authenticator = au
 }
 
 func ReCreateHTTP(port int) error {
